+++ calc/calc_multiply.go
@@ -1,6 +1,6 @@
 package calc
 
 // Multiply takes two numbers and adds them together
-func Multiply(a int, b int) int {
+func Multiply(a float32, b float32) float32 {
     return a*b;
 }
